requirements.txt
cassandra_driver==3.27.0
numpy==1.23.5
pandas==2.0.0
pytest==7.3.1
pytest-html==3.2.0
scikit_learn==1.2.2
